requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
colour==0.1.5
cryptography==46.0.2
Django==4.2.25
django-bootstrap-datepicker-plus==5.0.5
django-extensions==4.1
django-fileprovider==0.1.4
django-mathfilters==1.0.0
django-mongoengine==0.5.6
django-theme-soft-design==1.0.11
dnspython==2.7.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.11
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mongoengine==0.29.1
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
pymongo==4.15.3
pyshorteners==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-http-client==3.3.7
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sendgrid==6.12.5
six==1.17.0
sqlparse==0.5.3
str2bool==1.1
styleframe==4.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise==6.11.0
